poisson lambda type error

+++ ProbabilityMonad/Primitive/Poisson.cs
@@ -8,10 +8,10 @@ namespace ProbCSharp
     /// </summary>
     public class PoissonPrimitive : PrimitiveDist<int>
     {
-        public int Lambda;
+        public double Lambda;
         public Poisson Dist;
         public Random Gen;
-        public PoissonPrimitive(int lambda, Random gen)
+        public PoissonPrimitive(double lambda, Random gen)
         {
             Lambda = lambda;
             Dist = new Poisson(lambda);
